Add files via upload
This document contains a version of the NTC Billing Model used as the basis for the software that was used in production, world-wide by Nortel.  It is a very good example of the detail and sophistication required for operating at a World-Class level across multiple continents and multiple jurisdictions, including elements such as taxes,invoicing, business entities, agreements, agreement terms, reports, authorizations, changes, etc.