Add files via upload
Multi-sheet worksheet used for facilitating and justifying, as objectively as possible, which of the many servers and services provided by NORTEL's DtM (Design-to-Manufacturing) should see their Service Level Classifications downgraded or upgraded, as part of a cost-reduction initiative.  This was done in the lead up to the behind-the-scenes (not known to lower level employees) workup towards the outsourcing of all NORTEL Corporate IT Services to CSC.  NORTEL, circa 2001.